Implement sliding piece move generation
This is done via magic bitboards that are available at compile-time.
They will live in the program's binary. This reduces the startup time
since the tables don't need to be initialized during runtime.

This also allows for keeping the crate "no_std"-compatible since it
doesn't rely on incompatible dependencies such as the rand crate.