Revert 81a008598ee66ef899633f36e193104d8d0c9999
This reverts "Improve PieceMap::is_empty non-simd performance", which
lead to a small gain in a way that's totally up to the compiler. There
may be a bug in the future that makes that area perform worse.

Such a case is LLVM making the reference array into a static and calling
memcmp on it.